// frontend/tools/spectra/langium/ls/generated/ast.ts
export type Decl = Counter | Define | DefineRegExp | EXGar | LTLAsm | LTLGar | Monitor | Pattern | Predicate | RegexpTest | TypeDef | Var | WeightDef;

export const Decl = {
    $type: 'Decl'
} as const;
// ...
export interface DomainVarDecl extends langium.AstNode {
    readonly $container: QuantifierExpr;
    readonly $type: 'DomainVarDecl';
    domainType: VarType;
    name: string;
}

export const DomainVarDecl = {
    $type: 'DomainVarDecl',
    domainType: 'domainType',
    name: 'name'
} as const;
// ...
export interface Import extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Import';
    importURI: string;
}

export const Import = {
    $type: 'Import',
    importURI: 'importURI'
} as const;
// ...
export interface Model extends langium.AstNode {
    readonly $type: 'Model';
    elements: Array<Decl>;
    imports: Array<Import>;
    name: string;
}

export const Model = {
    $type: 'Model',
    elements: 'elements',
    imports: 'imports',
    name: 'name'
} as const;

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model.$type);
}

export interface Monitor extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Monitor' | 'TOK_SEMI';
    initial: Array<TemporalExpression>;
    name: string;
    safety: Array<TemporalExpression>;
    stateInv: Array<TemporalExpression>;
    type: VarType;
}

export const Monitor = {
    $type: 'Monitor',
    initial: 'initial',
    name: 'name',
    safety: 'safety',
    stateInv: 'stateInv',
    type: 'type'
} as const;
// ...
export interface Pattern extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Pattern' | 'TOK_SEMI';
    initial: Array<TemporalExpression>;
    justice: Array<TemporalExpression>;
    name: string;
    params?: PatternParamList;
    safety: Array<TemporalExpression>;
    stateInv: Array<TemporalExpression>;
    varDeclList: Array<VarDecl>;
}

export const Pattern = {
    $type: 'Pattern',
    initial: 'initial',
    justice: 'justice',
    name: 'name',
    params: 'params',
    safety: 'safety',
    stateInv: 'stateInv',
    varDeclList: 'varDeclList'
} as const;
// ...
export interface PatternParamList extends langium.AstNode {
    readonly $container: Pattern;
    readonly $type: 'PatternParamList';
    params: Array<PatternParam>;
}

export const PatternParamList = {
    $type: 'PatternParamList',
    params: 'params'
} as const;
// ...
export interface SizeDefineDecl extends langium.AstNode {
    readonly $container: DefineDecl | Subrange | VarType;
    readonly $type: 'SizeDefineDecl';
    arithExp?: TemporalExpression;
    sizeDef?: langium.Reference<DefineDecl>;
    value?: number;
}

export const SizeDefineDecl = {
    $type: 'SizeDefineDecl',
    arithExp: 'arithExp',
    sizeDef: 'sizeDef',
    value: 'value'
} as const;

export function isSizeDefineDecl(item: unknown): item is SizeDefineDecl {
    return reflection.isInstance(item, SizeDefineDecl.$type);
}

export interface Subrange extends langium.AstNode {
    readonly $container: Counter | VarType;
    readonly $type: 'Subrange';
    from: SizeDefineDecl;
    to: SizeDefineDecl;
}

export const Subrange = {
    $type: 'Subrange',
    from: 'from',
    to: 'to'
} as const;
// ...
export type TemporalExpression = Constant | QuantifierExpr | TemporalAdditiveExpr | TemporalAndExpr | TemporalBinaryExpr | TemporalIffExpr | TemporalImpExpr | TemporalInExpr | TemporalMultiplicativeExpr | TemporalOrExpr | TemporalPrimaryExpr | TemporalRelationalExpr | TemporalRemainderExpr | TemporalUnaryExpr;

export const TemporalExpression = {
    $type: 'TemporalExpression'
} as const;
// ...
export type TOK_SEMI = ';';
// ...
export interface TypeConstant extends langium.AstNode {
    readonly $container: VarType;
    readonly $type: 'TypeConstant';
    name: TypeConstantLiteral;
}

export const TypeConstant = {
    $type: 'TypeConstant',
    name: 'name'
} as const;

export function isTypeConstant(item: unknown): item is TypeConstant {
    return reflection.isInstance(item, TypeConstant.$type);
}

export type TypeConstantLiteral = number | string;


export interface TypeDef extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'TOK_SEMI' | 'TypeDef';
    name: string;
    type: VarType;
}

export const TypeDef = {
    $type: 'TypeDef',
    name: 'name',
    type: 'type'
} as const;

export function isTypeDef(item: unknown): item is TypeDef {
    return reflection.isInstance(item, TypeDef.$type);
}

export interface TypedParam extends langium.AstNode {
    readonly $container: TypedParamList;
    readonly $type: 'TypedParam';
    name: string;
    type: VarType;
}

export const TypedParam = {
    $type: 'TypedParam',
    name: 'name',
    type: 'type'
} as const;
// ...
export interface Var extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'TOK_SEMI' | 'Var';
    kind: VarOwner;
    name: string;
    type: VarType;
}

export const Var = {
    $type: 'Var',
    kind: 'kind',
    name: 'name',
    type: 'type'
} as const;

export function isVar(item: unknown): item is Var {
    return reflection.isInstance(item, Var.$type);
}

export interface VarDecl extends langium.AstNode {
    readonly $container: Pattern;
    readonly $type: 'TOK_SEMI' | 'VarDecl';
    name: string;
    type: VarType;
}

export const VarDecl = {
    $type: 'VarDecl',
    name: 'name',
    type: 'type'
} as const;

export function isVarDecl(item: unknown): item is VarDecl {
    return reflection.isInstance(item, VarDecl.$type);
}

export type VarOwner = 'aux' | 'auxvar' | 'env' | 'envvar' | 'in' | 'input' | 'out' | 'output' | 'sys' | 'sysvar';

export interface VarType extends langium.AstNode {
    readonly $container: DomainVarDecl | Monitor | TypeDef | TypedParam | Var | VarDecl;
    readonly $type: 'VarType';
    const: Array<TypeConstant>;
    dimensions: Array<SizeDefineDecl>;
    name?: 'boolean';
    subr?: Subrange;
    type?: langium.Reference<TypeDef>;
}

export const VarType = {
    $type: 'VarType',
    const: 'const',
    dimensions: 'dimensions',
    name: 'name',
    subr: 'subr',
    type: 'type'
} as const;